requirements.txt
pandas
psycopg2
SQLAlchemy
python-dotenv
